Fixed some UI issues, 2 components for each post.
