+++ 3Design/frontend/src/components/HomePage/HomePage.tsx
@@ -6,7 +6,7 @@ import { Button, Upload } from 'antd'
 import { Add, UploadOutlined } from '@mui/icons-material'
 import { Dialog, TextField } from '@mui/material'
 import CreatePost from '../CreatePost/CreatePost'
-import GalleryPost from '../GalleryPost/GalleryPost'
+import GalleryPost from '../GalleryPost/Clickable/GalleryPost'
 import Feed from '../Feed/Feed'
 import { useLocation, useParams } from 'react-router-dom'
 import GenericFeed from '../GenericFeed/GenericFeed'
