+++ 3Design/frontend/src/components/PostPage/PostPage.module.css
@@ -1,0 +1,6 @@
+.mainContainer{
+    width: 60%;
+    padding: 1rem;
+    display: flex;
+    flex-direction: column;
+}
