+++ 3Design/frontend/src/resources/json-files/MockDiscussions.json
@@ -1,6 +1,6 @@
 [
     {
-        "id": 5,
+        "id": 7,
         "title": "How do I print a model?",
         "body": "Which button should I press to start printing?",
         "memberId": 2,
@@ -16,7 +16,7 @@
         "bookmark":false
     },
     {
-        "id": 21,
+        "id": 23,
         "title": "How can I log out from this site?",
         "body": "I think they forgot to implented this feature.",
         "memberId": 2,
